+++ OVirtualMachine/Scenary.cpp
@@ -10,7 +10,7 @@
 
 namespace O{
 
-    std::pair<int, Scenary::ScriptWord> Scenary::generateScript(int *a, int id) {
+    std::pair<int, Scenary::ScriptWord> Scenary::generateScript(long *a, int id) {
         int readSize = 0;
         auto type = (InstructionType)a[id];
         readSize += 1;
@@ -32,7 +32,7 @@ namespace O{
         return {readSize, sw};
     }
 
-    std::pair<int, Scenary::Operand> Scenary::readOperand(Scenary::OperandType opT, int *a, int id) {
+    std::pair<int, Scenary::Operand> Scenary::readOperand(Scenary::OperandType opT, long *a, int id) {
         int readSize = 0;
         Scenary::Operand toRet;
         switch (opT) {
